Core/PacketIO: Updated SMSG_UPDATE_OBJECT

+++ src/server/game/Entities/Item/Item.cpp
@@ -1157,7 +1157,7 @@ void Item::BuildDynamicValuesUpdate(uint8 updateType, ByteBuffer* data, Player* 
 
             }
 
-            fieldBuffer << uint8(arrayMask.GetBlockCount());
+            fieldBuffer << uint16(arrayMask.GetBlockCount());
             arrayMask.AppendToPacket(&fieldBuffer);
             fieldBuffer.append(buffer);
         }
